Created AccordionPage.js
Refactored App.js to create new component - dropdown
